refactor(components): remove internal.ts, move scroll lock to utils/scroll-lock.ts
